move action to route file
