+++ app/routes/_app.inventory.$type/_index.tsx
@@ -1,6 +1,7 @@
 import * as RemixNode from '@remix-run/node'
 import * as RemixReact from '@remix-run/react'
 import * as React from 'react'
+import * as RemixParamsHelper from 'remix-params-helper'
 import invariant from 'tiny-invariant'
 import * as Zod from 'zod'
 import ItemList from '~/components/ItemList'
@@ -11,7 +12,27 @@ import * as InventoryModel from '~/models/inventory.server'
 import * as Session from '~/session.server'
 import * as Utils from '~/utils'
 
-export { action } from '~/actions/inventory'
+export async function action({ request }: RemixNode.LoaderArgs) {
+  const accountId = await Session.requireAccountId(request)
+
+  const ParamsSchema = Zod.object({
+    name: Zod.string(),
+    quantity: Zod.number().min(0).max(9999),
+  })
+
+  const { name, quantity } = await RemixParamsHelper.getFormDataOrFail(
+    request,
+    ParamsSchema
+  )
+
+  await InventoryModel.upsertInventoryItem({
+    accountId,
+    name,
+    quantity,
+  })
+
+  return null
+}
 
 interface LoaderData {
   items: ItemData.ItemWithQuantity[]
